refactor(combobox): separate into multiple files for multiple instance
