+++ packages/admin-ui/src/combobox/index.ts
@@ -2,3 +2,7 @@ export * from './combobox.state'
 export * from './combobox-item'
 export * from './combobox-field'
 export * from './combobox-popover'
+export * from './combobox-multiple-popover'
+export * from './combobox-multiple-item'
+export * from './combobox-multiple.state'
+export * from './combobox-multiple-field'
